apps/ml-service/pyproject.toml
[project]
name = "bmad-ml-service"
version = "0.1.0"
description = "ML/AI service for BMAD stock agent (FastAPI + OpenAI)."
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.8.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.1",
    "asyncpg>=0.29.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.9",
    "openai>=1.12.0",
    "mlflow>=2.10.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "psutil>=5.9.0",
    "apscheduler>=3.10.0",
]
[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "pytest-mock", "httpx", "black", "ruff"]
ml-torch = ["torch>=2.1.0"]
ml-tensorflow = ["tensorflow>=2.15.0"]
ml-all = ["torch>=2.1.0", "tensorflow>=2.15.0"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

